Added Patient Registration
Added a hard-coded patient registration form that populates patient db
as well as an xml to be sent to OpenMRS server. Registration xml and
patient will always be available to provider on phone, while the
registration form will help to identify new pts and whether they should
be assigned to a new provider.