Added/Rewrote A Background Download Service to run periodically, Changed DbHelper to Singleton
The download service now works again after another complete rewrite.
Should download based on specifications in the AlarmReceiver.  Alarm
called on power connected and on boot to reschedule alarms depending on
the db.  Since there is now service interacting with Db, also had to
make DbHelper a singleton.  There is one known ANR, which happens when
the service is downloading and clinic is open, and somehow clinic does
not close the RefreshDataActivity properly, so windowmanager loses
track of window and ANR.