internet searching, soul searching and a wide open database
a new ANR kept appearing due to multiple database connections. now the
Db access is left wide open... as in I never close the database! we
will see how it goes... but found good sources for this method.  so
far, so good.  it has even seemed to resolve my only other known ANR...
good to go.