Finally adding security! Move to Internal DBs, AES Encryption of SD, managed by wakelock alarm services
Major overhaul of all the security features, intended to work more
closely with the CHV Settings App also in progress. The move to
internal databases is meant to be more secure than external, and the
external databases are now using 256 bit encryption.  In addition, the
downloads now occur to internal store, in case there is corruption
during downloads or an FC, and they don't get deleted and remain open
on the SD.  Encryption is 256 bit symmetric, with keys stored internal
in db. Each form has unique key, and different ivs for each associated
form file (pics, audio etc.)  Encryption only occurs after submission
of the form. Wakelock alarms ensure that encryption and decryption
happens on a regular basis to ensure security of patient level data. Of
course, more work needed to improve the key security, but basically,
this is security aimed at lost or stolen SD cards.