Changing Collect Integration from ContentResolver() to SharedUserId
This is not a great design pattern, though it is used by Android system
apps to allow tight integration (and ease of coding).  Given that
Clinic is entirely dependent on Collect, this sharedUserId seems
somewhat justified though, and given how complicated the code base
would need to be to change form sharing resources for loading forms,
editing forms, saving completed forms, encrypting and decrypting forms,
viewing forms, and sending forms (which all occur across processes), it
seems more justifiable to do this.  In the end, given the
interdependency, this is probably the first step in merging the two
applications into one.