Fixed bug with SSL Certificate updates
Fixed bug with the recognition of aliases within the truststore for
truststores that were not made through keytool with defined hash
aliases.  Now works with truststores regardless of how they were made
or put onto the phone, and so should be able to update certificates
dynamically for all bouncycastle truststores.