Adding Feature to request Patient Consent with Signature in order to view OpenMRS data
This is intended to provide CDS Smart Form capability and share data
across providers, while still ensuring patient rights are observed.
This feature allows CHVs to fill out forms as usual, but gives option
to view the data and the patient-specific SmartForms on that patient
downloaded from OpenMRS if they obtain consent. This is an optional
feature, with settings added in the menu preference to turn it off.
Signatures are stored as ~10k blobs with consent 'watermark' of client
name and ID, Provider ID, Date and Consent duration in a new database
table 'Consent.'  Previous consent can be viewed and also removed at
any time. Db keeps track of old consents for a period of time
(specified in prefs) to prevent spurious consenting without the patient
present.  New consent is requested on each ViewPatient until patient
either consents or does not consent.  Consent Form is simply a webpage
(no javascript) which allows for implementation-specific consents by
simply adding the form to the sd card; if no form is available, a
default consent form is provided in Assets.  Code is cleaned-up to
provide for these new activities with reorganization of some
inheritance and utilities.  Safely does not apply for patients with no
OpenMRS data (registered on phone)