Fixed consent loophole
Added consent expiration date to data model so that the consent would
be for the time period specified at the time of signature, rather than
the consent period specified by the time on the phone.  This allows
client consent to be secure only for the specified period that they
signed onto, regardless of what other settings may be set later.